% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll-call.r
\name{fortify.pprc}
\alias{fortify.pprc}
\title{Fortify a \code{\link[=roll_call]{roll_call()}} (\code{pprc}) object}
\usage{
\method{fortify}{pprc}(model, data, ...)
}
\arguments{
\item{model}{a \code{\link[=roll_call]{roll_call()}} (\code{pprc}) object}

\item{data}{unused}

\item{...}{unused}
}
\description{
Fortify a \code{\link[=roll_call]{roll_call()}} (\code{pprc}) object
}
