% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghouse.r
\name{house_carto}
\alias{house_carto}
\title{Produce a House cartogram}
\usage{
house_carto(vote_tally)
}
\arguments{
\item{vote_tally}{either a \code{pprc} object (the result of a call to \code{\link[=roll_call]{roll_call()}}) or
a \code{data.frame} of vote tallys for the house It expects 3 columns. \code{state_abbrev} : the
2-letter U.S. state abbreviation; \code{district} : either \code{1} or \code{2} to distinguish between
each representative; \code{party} : \code{R}, \code{D} or \code{ID}; \code{position} : \code{yes}, \code{no}, \code{present}, \code{none} for
how the representative voted.}
}
\value{
a \code{ggplot2} object that you can further customize with scales, labels, etc.
}
\description{
Produce a House cartogram
}
