% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll-call.r
\name{print.pprc}
\alias{print.pprc}
\title{Better default \code{print} function for \code{\link[=roll_call]{roll_call()}} (\code{pprc}) objects}
\usage{
\method{print}{pprc}(x, ...)
}
\arguments{
\item{x}{a \code{\link[=roll_call]{roll_call()}} (\code{pprc}) object}

\item{...}{ignored}
}
\description{
Better default \code{print} function for \code{\link[=roll_call]{roll_call()}} (\code{pprc}) objects
}
