% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll-call.r
\name{roll_call}
\alias{roll_call}
\title{Get Voting Record for House or Senate By Number, Session & Roll Call Number}
\usage{
roll_call(critter = c("house", "senate"), number, session = c(1L, 2L),
  rcall)
}
\arguments{
\item{critter}{one of `house` or `senate`}

\item{number}{valid congress number (e.g. `115`)}

\item{session}{a valid session numbner (i.e. `1` or `2` and valid for current year)}

\item{rcall}{roll call vote number}
}
\description{
Get Voting Record for House or Senate By Number, Session & Roll Call Number
}
