% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.r
\name{theme_voteogram}
\alias{theme_voteogram}
\title{voteogram ggplot2 theme}
\usage{
theme_voteogram(legend = TRUE)
}
\arguments{
\item{legend}{if `FALSE` no legend is shown}
}
\description{
voteogram ggplot2 theme
}
