% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-colors.r
\docType{data}
\name{voteogram-colors}
\alias{voteogram-colors}
\alias{vote_carto_fill}
\alias{vote_carto_color}
\title{voteogram manual scale colors}
\format{
An object of class \code{character} of length 13.

An object of class \code{character} of length 13.
}
\usage{
vote_carto_fill

vote_carto_color
}
\description{
Manual color scales fill and color scale values (in the event you need to use them)
outisde the main plotting functions.
}
\keyword{datasets}
