% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voteogram-package.R
\name{voteogram-exports}
\alias{voteogram-exports}
\alias{\%>\%}
\title{voteogram exported operators}
\description{
The following functions are imported and then re-exported
from the voteogram package to enable use of the magrittr
pipe operator with no additional library calls
}
