% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voteogram-package.R
\docType{package}
\name{voteogram}
\alias{voteogram}
\alias{voteogram-package}
\title{U.S. House and Senate Voting Cartogram Generators}
\description{
Produce voting cartograms in the style of 'ProPublica'.
}
\author{
Bob Rudis (bob@rud.is)
}
